Separate BOX and Bounds.
