+++ pdal/PointView.hpp
@@ -41,7 +41,6 @@
 #include <pdal/PointLayout.hpp>
 #include <pdal/PointRef.hpp>
 #include <pdal/PointTable.hpp>
-#include <pdal/util/Bounds.hpp>
 
 #include <memory>
 #include <queue>
@@ -62,6 +61,8 @@ class PointView;
 class PointViewIter;
 class KD2Index;
 class KD3Index;
+class BOX2D;
+class BOX3D;
 
 typedef std::shared_ptr<PointView> PointViewPtr;
 typedef std::set<PointViewPtr, PointViewLess> PointViewSet;
