+++ plugins/i3s/io/EsriReader.hpp
@@ -41,8 +41,8 @@
 #include <json/json.h>
 #include <arbiter/arbiter.hpp>
 
+#include <pdal/Bounds.hpp>
 #include <pdal/Reader.hpp>
-#include <pdal/util/Bounds.hpp>
 #include <pdal/util/IStream.hpp>
 
 namespace pdal
