+++ test/unit/BoundsTest.cpp
@@ -33,7 +33,7 @@
 ****************************************************************************/
 
 #include <pdal/pdal_test_main.hpp>
-#include <pdal/util/Bounds.hpp>
+#include <pdal/Bounds.hpp>
 #include <pdal/PDALUtils.hpp>
 
 using namespace pdal;
